Implements a first try at the build structure.
A test fails, but since I can't really figure out how to use GDB on OSX,
I wanted to push it in and clone it on another machine.

All code that was previously in main.cpp was split up into seperate
files that each contain a class implementation under /src.

Adds a CMake configuration for compiling the renderer as a shared
library.

Adds Catch to the project as a testing framework.

Also builds a Catch testrunner executable that links to the output of
the compilation of the renderer.

Some simple tests were implemented to verify that the structure was
working as I wanted.