Improves GLContextParam comparisons and tests
Adds a test for sorting a vector of GLContextParams using STL's
stable_sort. Initially this didn't have a desired output, which led to
this commit.

Comparisons are now weighed. The versions are now the most important property,
and the rest of the properties only matter if the versions are equal.